+++ frontend/app/src/components/status/sync/SyncIndicator.vue
@@ -14,13 +14,15 @@ const {
   confirmChecked,
   displaySyncConfirmation,
   syncAction,
+  uploadStatus,
   cancelSync,
   forceSync,
   showSyncConfirmation,
+  clearUploadStatus,
 } = useSync();
 const { href, onLinkClick } = useLinks();
 
-const premium = usePremium();
+const { premium, premiumSync } = storeToRefs(usePremiumStore());
 
 const pending = ref<boolean>(false);
 const visible = ref<boolean>(false);
@@ -48,12 +50,33 @@ const icon = computed(() => {
   return tick ? 'upload-cloud-2-line' : 'upload-cloud-line';
 });
 
+const showAutoUploadOffWarning = computed<boolean>(() => {
+  if (!isDefined(uploadStatus))
+    return false;
+
+  return !get(uploadStatus).uploaded;
+});
+
+const tooltip = computed<string>(() => {
+  if (get(showAutoUploadOffWarning)) {
+    const title = t('sync_indicator.db_upload_result.title');
+    const message = t('sync_indicator.db_upload_result.message', {
+      reason: get(uploadStatus)?.message,
+    });
+    return `${title}: ${message}`;
+  }
+  return t('sync_indicator.menu_tooltip');
+});
+
 function showConfirmation(action: SyncAction) {
   set(visible, false);
   showSyncConfirmation(action);
 }
 
 async function performSync() {
+  if (get(syncAction) === SYNC_UPLOAD)
+    clearUploadStatus();
+
   resume();
   set(pending, true);
   await forceSync(logout);
@@ -83,19 +106,38 @@ watch(isSyncing, (current, prev) => {
       >
         <template #activator="{ on }">
           <MenuTooltipButton
-            :tooltip="t('sync_indicator.menu_tooltip')"
+            :tooltip="tooltip"
             class-name="secondary--text text--lighten-4"
             v-on="on"
           >
-            <RuiIcon
-              v-if="isSyncing"
-              :name="icon"
-              color="primary"
-            />
-            <RuiIcon
-              v-else
-              name="cloud-line"
-            />
+            <RuiBadge
+              :value="showAutoUploadOffWarning"
+              color="warning"
+              dot
+              placement="top"
+              offset-y="4"
+              size="lg"
+              class="flex items-center"
+            >
+              <RuiIcon
+                v-if="showAutoUploadOffWarning"
+                name="cloud-off-line"
+                color="warning"
+              />
+              <RuiIcon
+                v-else-if="!premiumSync"
+                name="cloud-off-line"
+              />
+              <RuiIcon
+                v-else-if="isSyncing"
+                :name="icon"
+                color="primary"
+              />
+              <RuiIcon
+                v-else
+                name="cloud-line"
+              />
+            </RuiBadge>
           </MenuTooltipButton>
         </template>
         <div class="pa-4 md:w-[250px] w-full">
@@ -110,6 +152,32 @@ watch(isSyncing, (current, prev) => {
             <span v-else>
               {{ t('common.never') }}
             </span>
+          </div>
+          <div
+            v-if="uploadStatus"
+            class="flex flex-col my-2 p-2 gap-2 border border-rui-warning rounded-[0.25rem]"
+          >
+            <div class="flex gap-1">
+              <div class="font-medium">
+                {{ t('sync_indicator.db_upload_result.title') }}
+              </div>
+              <RuiButton
+                variant="text"
+                icon
+                size="sm"
+                @click="clearUploadStatus()"
+              >
+                <RuiIcon name="close-line" />
+              </RuiButton>
+            </div>
+
+            <div class="text-secondary">
+              {{
+                t('sync_indicator.db_upload_result.message', {
+                  reason: uploadStatus.message,
+                })
+              }}
+            </div>
           </div>
           <SyncButtons
             :pending="pending"