+++ rotkehlchen/db/history_events.py
@@ -8,7 +8,16 @@ from pysqlcipher3 import dbapi2 as sqlcipher
 from rotkehlchen.assets.asset import Asset
 from rotkehlchen.constants import ZERO
 from rotkehlchen.constants.limits import FREE_HISTORY_EVENTS_LIMIT
-from rotkehlchen.db.constants import HISTORY_MAPPING_KEY_STATE, HISTORY_MAPPING_STATE_CUSTOMIZED
+from rotkehlchen.db.constants import (
+    ETH_STAKING_EVENT_FIELDS,
+    ETH_STAKING_FIELD_LENGTH,
+    EVM_EVENT_FIELDS,
+    EVM_FIELD_LENGTH,
+    HISTORY_BASE_ENTRY_FIELDS,
+    HISTORY_BASE_ENTRY_LENGTH,
+    HISTORY_MAPPING_KEY_STATE,
+    HISTORY_MAPPING_STATE_CUSTOMIZED,
+)
 from rotkehlchen.db.filtering import (
     ALL_EVENTS_DATA_JOIN,
     EVM_EVENT_JOIN,
@@ -54,14 +63,17 @@ logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 
-HISTORY_BASE_ENTRY_FIELDS = 'entry_type, history_events.identifier, event_identifier, sequence_index, timestamp, location, location_label, asset, amount, usd_value, notes, type, subtype '  # noqa: E501
-HISTORY_BASE_ENTRY_LENGTH = 12
-
-EVM_EVENT_FIELDS = 'tx_hash, counterparty, product, address, extra_data'
-EVM_FIELD_LENGTH = 5
-
-ETH_STAKING_EVENT_FIELDS = 'validator_index, is_exit_or_blocknumber'
-ETH_STAKING_FIELD_LENGTH = 2
+def maybe_filter_ignore_asset(
+        filter_query: HistoryBaseEntryFilterQuery,
+) -> str:
+    """An auxiliar function to find if query_filter contains `DBIgnoredAssetsFilter`. If it does
+    then return that filter clause. This is done where we want to filter ignored assets
+    before applying the free limit."""
+    for fil in filter_query.filters:
+        if isinstance(fil, DBIgnoredAssetsFilter):
+            # Also don't count spam asset transactions in the limit
+            return 'WHERE (asset IS NULL OR asset NOT IN (SELECT value FROM multisettings WHERE name="ignored_asset")) '  # noqa: E501
+    return ' '
 
 
 class DBHistoryEvents:
@@ -354,36 +366,37 @@ class DBHistoryEvents:
         TODO: To not query all columns with all joins for all cases, we perhaps can
         peek on the entry type of the filter and adjust the SELECT fields accordingly?
         """
-        free_query_group_by = ''
-        free_query_count = ''
         base_prefix = 'SELECT '
-        type_idx = 0
+        base_suffix = f'{HISTORY_BASE_ENTRY_FIELDS}, {EVM_EVENT_FIELDS}, {ETH_STAKING_EVENT_FIELDS} {ALL_EVENTS_DATA_JOIN}'  # noqa: E501
+        bindings = []
+        type_idx = 1 if group_by_event_ids is True else 0
         special_free_query = False
-        if group_by_event_ids is True:
-            if has_premium:
+        free_query_group_by = maybe_filter_ignore_asset(filter_query)
+
+        if has_premium is True:
+            if group_by_event_ids is True:
                 base_prefix += 'COUNT(*), '
-            else:  # a bit ugly conditions to keep limit at groups for free users
-                free_query_count = 'COUNT(*), '
-                for fil in filter_query.filters:
-                    if isinstance(fil, DBIgnoredAssetsFilter):
-                        # Also don't count spam asset transactions in the limit
-                        free_query_group_by = 'WHERE (asset IS NULL OR asset NOT IN (SELECT value FROM multisettings WHERE name="ignored_asset")) '  # noqa: E501
-                        break
+            base_query = f'{base_prefix} {base_suffix}'
+        else:
+            if group_by_event_ids is True:
+                special_free_query = True  # a bit ugly conditions to keep limit at groups for free users  # noqa: E501
                 free_query_group_by += 'GROUP BY event_identifier'
-                special_free_query = True
+                base_query = (
+                    f'{base_prefix} * FROM (SELECT COUNT(*), {base_suffix} '
+                    f'{free_query_group_by} ORDER BY timestamp DESC, sequence_index ASC LIMIT ?) '
+                )
+            else:  # if we don't group them but still apply free limit, then apply this limit only on any event_identifiers that fall inside the limit  # noqa: E501
+                base_query = (
+                    f'{base_prefix} * FROM (SELECT {base_suffix} WHERE event_identifier IN '
+                    f'(SELECT DISTINCT event_identifier FROM history_events {free_query_group_by} ORDER BY timestamp DESC, sequence_index ASC LIMIT ?)) '  # noqa: E501
+                )
+            bindings = [FREE_HISTORY_EVENTS_LIMIT]
 
-            type_idx = 1
-
-        prepared_query, bindings = filter_query.prepare(
+        prepared_query, prepared_bindings = filter_query.prepare(
             with_group_by=group_by_event_ids,
             special_free_query=special_free_query,
         )
-
-        if has_premium is True:
-            base_query = f'{base_prefix} {HISTORY_BASE_ENTRY_FIELDS}, {EVM_EVENT_FIELDS}, {ETH_STAKING_EVENT_FIELDS} {ALL_EVENTS_DATA_JOIN}'  # noqa: E501
-        else:
-            base_query = f'{base_prefix} * FROM (SELECT {free_query_count} {HISTORY_BASE_ENTRY_FIELDS}, {EVM_EVENT_FIELDS}, {ETH_STAKING_EVENT_FIELDS} {ALL_EVENTS_DATA_JOIN} {free_query_group_by} ORDER BY timestamp DESC, sequence_index ASC LIMIT ?) '  # noqa: E501
-            bindings.insert(0, FREE_HISTORY_EVENTS_LIMIT)
+        bindings.extend(prepared_bindings)
 
         cursor.execute(base_query + prepared_query, bindings)
         output: list[HistoryBaseEntry] | list[tuple[int, HistoryBaseEntry]] = []
@@ -600,21 +613,28 @@ class DBHistoryEvents:
         """
         prepared_query, bindings = query_filter.prepare(with_pagination=False)
         # we need to select everything because any column could be used in the filter
-        query = 'SELECT * ' + query_filter.get_join_query() + prepared_query
+        query = f'SELECT {query_filter.get_columns()} {query_filter.get_join_query()} {prepared_query}'  # noqa: E501
         if group_by_event_ids:
             query = f'SELECT event_identifier FROM ({query}) GROUP BY event_identifier'
         query = f'SELECT COUNT(*) FROM ({query})'
         count_without_limit = cursor.execute(query, bindings).fetchone()[0]
 
         if entries_limit is not None:
-            query = 'SELECT * ' + query_filter.get_join_query()
-            if group_by_event_ids:  # we take the groups before the limit has been applied
-                query += ' GROUP BY event_identifier '
-            query += ' ORDER BY timestamp DESC LIMIT ?'
-            bindings.insert(0, entries_limit)
-            query = f'SELECT COUNT(*) FROM ({query}) ' + prepared_query
+            prepared_query, bindings = query_filter.prepare(
+                with_group_by=False,
+                special_free_query=True,
+                with_order=False,
+                with_pagination=False,
+            )
 
-            count_with_limit = cursor.execute(query, bindings).fetchone()[0]
+            free_query_group_by = maybe_filter_ignore_asset(query_filter)
+            count_with_limit = cursor.execute(
+                f'SELECT COUNT(*) FROM ('
+                f'SELECT {query_filter.get_columns()} {query_filter.get_join_query()}{free_query_group_by}'  # we take the groups before the limit has been applied  # noqa: E501
+                'GROUP BY event_identifier ORDER BY timestamp DESC, sequence_index ASC LIMIT ?'
+                f'){prepared_query}',
+                [entries_limit] + bindings,  # add limit's binding before prepared_query's bindings
+            ).fetchone()[0]
             return count_without_limit, count_with_limit
 
         return count_without_limit, count_without_limit